Add files via upload
This is a cylinder that I rigged, it moves like a worm! 
I subdivided the cylinder and added bones to it. I then attached it to it's skin (the cylinder) and then created controls (the circles) and constraints for animators to use. 
I struggled for a second because I didn't select my joints and circles in the correct order, so I had to go back and try again!